Add unlock instruction to all pause checks
When the ERCF is locked, it's easy (if you're new at least) to forget to
unlock it.  When this happens and you're trying to fix an issue, all the
commands silently return with no response.  This changes all of the
pause checks to print out a message about running unlock.  This might
lead to the unlock message being printed more than once in some cases,
but that's probably better than not at all.